simplify main.go
